requirements.txt
llama-index>=0.9.8
openai>=1.8.0
python-dotenv>=1.0.0
composio>=0.1.0
composio-llamaindex>=0.1.0
apscheduler>=3.10.4
pyyaml>=6.0.1
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0 